+++ crates/amaru-ledger/src/store/columns/mod.rs
@@ -25,6 +25,7 @@ pub mod recently_unregistered_accounts;
 pub mod slots;
 pub mod utxo;
 pub mod votes;
+pub mod vrf_keys;
 
 #[expect(clippy::panic)]
 pub fn unsafe_decode<T: for<'d> cbor::Decode<'d, ()>>(bytes: &[u8]) -> T {